# Openapi.yaml
openapi: 3.1.0

info:
  title: Vehicle Tracking API
  description: API for managing fleet vehicles, sensor readings, and logs.
  version: 1.0.0

servers:
  - url: http://localhost
    description: Local Development Server

paths:
  /health:
    get:
      summary: Check API status
      responses:
        "200":
          description: OK

  /api/versions:
    get:
      summary: List API versions
      responses:
        "200":
          description: Versions list
        "429":
          $ref: '#/components/responses/TooManyRequests'

  # V1 vehicles
  /api/v1/vehicles:
    get:
      summary: List vehicles (v1)
      security:
        - ApiKeyAuth: []
      responses:
        "200":
          description: List of vehicles
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/VehicleV1'
        "429":
          $ref: '#/components/responses/TooManyRequests'

    post:
      summary: Create vehicle (v1)
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VehicleCreate'
      responses:
        "201":
          description: Vehicle Created
        "400":
          $ref: '#/components/responses/BadRequest'
        "409":
          description: Conflict
        "500":
          $ref: '#/components/responses/ServerError'

  /api/v1/vehicles/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
    get:
      summary: Get vehicle by ID (v1)
      security:
        - ApiKeyAuth: []
      responses:
        "200":
          description: Found Vehicle
        "404":
          $ref: '#/components/responses/NotFound'

    put:
      summary: Update vehicle (v1)
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VehicleUpdate'
      responses:
        "200":
          description: Vehicle Updated
        "404":
          $ref: '#/components/responses/NotFound'

    delete:
      summary: Delete vehicle (v1)
      security:
        - ApiKeyAuth: []
      responses:
        "204":
          description: No Content
        "404":
          $ref: '#/components/responses/NotFound'

  # V1 position
  /api/v1/vehicles/{id}/positions:
    post:
      summary: Log a vehicle position
      security:
        - ApiKeyAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PositionCreate'
      responses:
        "201":
          description: Position Logged
        "404":
          $ref: '#/components/responses/NotFound'

  # V1 fuel
  /api/v1/vehicles/{id}/fuel:
    post:
      summary: Log vehicle fuel level
      security:
        - ApiKeyAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FuelLogCreate'
      responses:
        "201":
          description: Fuel Logged
        "404":
          $ref: '#/components/responses/NotFound'

  # V2 vehicle
  /api/v2/vehicles:
    get:
      summary: List vehicles (v2)
      security:
        - BearerAuth: []
      responses:
        "200":
          description: List of vehicles
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/VehicleV2'

    post:
      summary: Create vehicle (v2)
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VehicleCreate'
      responses:
        "201":
          description: Vehicle Created
        "400":
          $ref: '#/components/responses/BadRequest'
        "409":
          description: Conflict

  /api/v2/vehicles/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
    get:
      summary: Get vehicle by ID (v2)
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Found Vehicle
        "404":
          $ref: '#/components/responses/NotFound'

    put:
      summary: Update vehicle (v2)
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VehicleUpdate'
      responses:
        "200":
          description: Vehicle Updated
        "404":
          $ref: '#/components/responses/NotFound'

    delete:
      summary: Delete vehicle (v2)
      security:
        - BearerAuth: []
      responses:
        "204":
          description: No Content
        "404":
          $ref: '#/components/responses/NotFound'

components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-KEY
    BearerAuth:
      type: http
      scheme: bearer

  responses:
    BadRequest:
      description: Bad Request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorDetail'
    NotFound:
      description: Resource Not Found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorDetail'
    TooManyRequests:
      description: Rate Limit Exceeded
      headers:
        X-RateLimit-Limit:
          schema:
            type: integer
        X-RateLimit-Remaining:
          schema:
            type: integer
        X-RateLimit-Reset:
          schema:
            type: integer
    ServerError:
      description: Internal Server Error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorDetail'

  schemas:
    ErrorDetail:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
        timestamp:
          type: string
          format: date-time

    VehicleCreate:
      type: object
      required:
        - plate
        - model
      properties:
        plate:
          type: string
        model:
          type: string
        driver:
          type: string
        status:
          type: string
          enum: [active, inactive, maintenance]

    VehicleUpdate:
      type: object
      properties:
        plate:
          type: string
        model:
          type: string
        driver:
          type: string
        status:
          type: string

    VehicleV1:
      type: object
      properties:
        id:
          type: integer
        plate:
          type: string
        model:
          type: string
        status:
          type: string
        _links:
          type: object

    VehicleV2:
      type: object
      properties:
        id:
          type: integer
        registration:
          type: string
        model:
          type: string
        status:
          type: string
        _links:
          type: object

    PositionCreate:
      type: object
      required:
        - latitude
        - longitude
      properties:
        latitude:
          type: number
        longitude:
          type: number
        speed_kmh:
          type: number

    FuelLogCreate:
      type: object
      required:
        - fuel_level
      properties:
        fuel_level:
          type: number
        fuel_litres:
          type: number